+++ CHANGELOG.md
@@ -34,6 +34,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Improving consistency with `git` handling of `.gitignore`, including `/`
   after entries and `\r\r\n` for MacOS files.
 
+* `jj status` filters untracked paths by fileset
+  [#9287](https://github.com/jj-vcs/jj/issues/9287)
+
 ## [0.40.0] - 2026-04-01
 
 ### Release highlights